refactor(tui): P12.4 extract stream arm + relocate handlers + LOC ratchet — WAVE 4 COMPLETE
Engine event_queue arm + in-loop stream .next() -> stream_handler.rs; P12.3 handler fns
-> loop_arms.rs. mod.rs 1279->431 (beats the ~600 goal). loc-ratchet.sh + loc_ratchet.rs
(ceiling 460) wired to CI so run() can't silently regrow. Streaming lifecycle byte-identical:
38 harness incl streaming, 331 unit, clippy 0, both ratchets green.
⚠️ HUMAN GATE P12.4: Haseeb live-drives streaming before merge (restart deref sites
stream_handler.rs:263-266/340-344/354-357). run() split = T125 (leave open for Haseeb).
