+++ .github/workflows/ci.yml
@@ -29,3 +29,9 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - run: bash scripts/ignore-ratchet.sh
+
+  loc-ratchet:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - run: bash scripts/loc-ratchet.sh
